Extracted nested classes PatternComparer, RecognizedChar and RecognizedPattern. They are no longer nested inside CharacterRecognizer but have their own files.
