+++ src/main/java/net/sf/javaanpr/recognizer/CharacterRecognizer.java
@@ -18,12 +18,6 @@ package net.sf.javaanpr.recognizer;
 
 import net.sf.javaanpr.imageanalysis.Char;
 
-import java.awt.*;
-import java.awt.image.BufferedImage;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.Vector;
-
 public abstract class CharacterRecognizer {
 
     public static final char[] ALPHABET =
@@ -45,120 +39,4 @@ public abstract class CharacterRecognizer {
     };
 
     public abstract RecognizedChar recognize(Char chr);
-
-    public static class RecognizedChar {
-        private Vector<RecognizedPattern> patterns;
-        private boolean isSorted;
-
-        public RecognizedChar() {
-            this.patterns = new Vector<RecognizedPattern>();
-            this.isSorted = false;
-        }
-
-        public void addPattern(RecognizedPattern pattern) {
-            this.patterns.add(pattern);
-        }
-
-        public boolean isSorted() {
-            return this.isSorted;
-        }
-
-        public void sort(int direction) {
-            if (this.isSorted) {
-                return;
-            }
-            this.isSorted = true;
-            Collections.sort(this.patterns, new PatternComparer(direction));
-        }
-
-        /**
-         * @return null if not sorted
-         */
-        public Vector<RecognizedPattern> getPatterns() {
-            if (this.isSorted) {
-                return this.patterns;
-            }
-            return null;
-        }
-
-        public RecognizedPattern getPattern(int i) {
-            if (this.isSorted) {
-                return this.patterns.elementAt(i);
-            }
-            return null;
-        }
-
-        public BufferedImage render() {
-            int width = 500;
-            int height = 200;
-            BufferedImage histogram = new BufferedImage(width + 20, height + 20, BufferedImage.TYPE_INT_RGB);
-            Graphics2D graphic = histogram.createGraphics();
-            graphic.setColor(Color.LIGHT_GRAY);
-            Rectangle backRect = new Rectangle(0, 0, width + 20, height + 20);
-            graphic.fill(backRect);
-            graphic.draw(backRect);
-            graphic.setColor(Color.BLACK);
-
-            int colWidth = width / this.patterns.size();
-            int left, top;
-            for (int ay = 0; ay <= 100; ay += 10) {
-                int y = 15 + (int) (((100 - ay) / 100.0f) * (height - 20));
-                graphic.drawString(new Integer(ay).toString(), 3, y + 11);
-                graphic.drawLine(25, y + 5, 35, y + 5);
-            }
-            graphic.drawLine(35, 19, 35, height);
-            graphic.setColor(Color.BLUE);
-            for (int i = 0; i < this.patterns.size(); i++) {
-                left = (i * colWidth) + 42;
-                top = height - (int) (this.patterns.elementAt(i).cost * (height - 20));
-                graphic.drawRect(left, top, colWidth - 2, height - top);
-                graphic.drawString(this.patterns.elementAt(i).chr + " ", left + 2, top - 8);
-            }
-            return histogram;
-        }
-
-        public static final class RecognizedPattern {
-            private char chr;
-            private float cost;
-
-            public RecognizedPattern(char chr, float value) {
-                this.chr = chr;
-                this.cost = value;
-            }
-
-            public char getChar() {
-                return this.chr;
-            }
-
-            public float getCost() {
-                return this.cost;
-            }
-        }
-
-        public static class PatternComparer implements Comparator<Object> {
-
-            private int direction;
-
-            public PatternComparer(int direction) {
-                this.direction = direction;
-            }
-
-            @Override
-            public int compare(Object o1, Object o2) {
-                float cost1 = ((RecognizedPattern) o1).getCost();
-                float cost2 = ((RecognizedPattern) o2).getCost();
-                int ret = 0;
-                if (cost1 < cost2) {
-                    ret = -1;
-                }
-                if (cost1 > cost2) {
-                    ret = 1;
-                }
-                if (this.direction == 1) {
-                    ret *= -1;
-                }
-                return ret;
-            }
-        }
-    }
 }
