+++ src/main/java/net/sf/javaanpr/recognizer/RecognizedPattern.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2013 JavaANPR contributors
+ * Copyright 2006 Ondrej Martinsky
+ * Licensed under the Educational Community License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.osedu.org/licenses/ECL-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an "AS IS"
+ * BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+package net.sf.javaanpr.recognizer;
+
+public final class RecognizedPattern {
+    private char chr;
+    private float cost;
+
+    public RecognizedPattern(char chr, float value) {
+        this.chr = chr;
+        this.cost = value;
+    }
+
+    public char getChar() {
+        return this.chr;
+    }
+
+    public float getCost() {
+        return this.cost;
+    }
+}
